Changes made to QuestionRepository.java

+++ src/main/java/eu/programit/repository/QuestionRepository.java
@@ -1,12 +1,25 @@
 package eu.programit.repository;
 
-import eu.programit.domain.Question;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Component;
 
-/**
- * Created by udr013 on 8-6-2016.
- */
+import eu.programit.domain.Question;
+
 @Component
-public interface QuestionRepository extends CrudRepository<Question,Integer> {
+public interface QuestionRepository extends CrudRepository<Question, Long> { 
+	
 }
+
+
+//package eu.programit.repository;
+//
+//import eu.programit.domain.Question;
+//import org.springframework.data.repository.CrudRepository;
+//import org.springframework.stereotype.Component;
+//
+///**
+//* Created by udr013 on 8-6-2016.
+//*/
+//@Component
+//public interface QuestionRepository extends CrudRepository<Question,Integer> {
+//}
